Refactor FieldMeta deserialization. Added support of declare custom widget fields in client projects
