+++ tesler-core/src/main/java/io/tesler/core/ui/model/json/SqlBindMeta.java
@@ -22,6 +22,7 @@ package io.tesler.core.ui.model.json;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+import io.tesler.core.ui.model.json.field.FieldMeta;
 import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
