+++ tesler-core/src/main/java/io/tesler/core/ui/model/json/field/subtypes/InputFieldMeta.java
@@ -18,32 +18,34 @@
  * #L%
  */
 
-package io.tesler.core.ui.model.json;
+package io.tesler.core.ui.model.json.field.subtypes;
 
-import io.tesler.core.ui.model.json.FieldMeta.FieldContainer;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import java.util.List;
+import io.tesler.core.ui.field.TeslerWidgetField;
+import io.tesler.core.ui.field.link.LinkToField;
+import io.tesler.core.ui.model.json.field.FieldMeta.FieldMetaBase;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.Setter;
 
-@Getter
-@Setter
-public final class FieldGroup implements FieldContainer {
+@EqualsAndHashCode(callSuper = true)
+@Data
+@TeslerWidgetField({"input", "hint"})
+public class InputFieldMeta extends FieldMetaBase {
 
-	@JsonProperty(value = "break")
-	private Boolean breakBlock;
+	@LinkToField
+	private String maskField;
 
-	private Boolean fullSize;
+	private InputMaskMeta mask;
 
-	private String blockId;
+	@Getter
+	@Setter
+	public static class InputMaskMeta {
 
-	private String name;
+		private String type;
 
-	private Boolean visible;
+		private String mask;
 
-	private Boolean newRow;
-
-	@JsonProperty("fields")
-	private List<FieldMeta> children;
+	}
 
 }
